Added FontManager and FontFactory to better manage fonts. Made default font use "C:\\Windows\\Fonts\\Arial.ttf" but permission is denied.

+++ Bolt-Core/src/Core/Math/Random/Random.cpp
@@ -13,7 +13,7 @@ namespace Bolt
 		m_Gen.seed((int)std::time(nullptr));
 	}
 
-	void Random::SetSeed(int64 seed)
+	void Random::SetSeed(uint seed)
 	{
 		m_Gen.seed(seed);
 	}
